requirements.txt
nose==1.3.7
numpy==1.16.3
Pillow==6.0.0
